Extract hardcoded i18n behaviour into re-usable utilities (3)
