+++ src/types/GraphQLDataProvider.ts
@@ -1,3 +1,7 @@
+/**
+ * Defined in node_modules/ra-data-graphql-prisma/typings.d.ts
+ * But aren't exported and must be re-declared here to be usable in our project
+ */
 export type GraphQLDataProvider = (
   fetchType: string,
   resource: string,
